requirements.txt
streamlit
Pandas
BeautifulSoup4
lxml
yfinance
matplotlib.pyplot
matplotlib.ticker
